Add Light component
