+++ src/engine/app.rs
@@ -7,7 +7,7 @@ use context::{Config, Context};
 use ecs::{Planner, World, Processor, Priority};
 use std::sync::{Arc, Mutex};
 use std::ops::DerefMut;
-use processors::Renderable;
+use processors::{Renderable, Light};
 
 /// User-friendly facade for building games. Manages main loop.
 pub struct Application {
@@ -99,6 +99,7 @@ impl<T> ApplicationBuilder<T>
     pub fn new(initial_state: T, config: Config) -> ApplicationBuilder<T> {
         let mut world = World::new();
         world.register::<Renderable>();
+        world.register::<Light>();
         let planner = Planner::new(world, 1);
         ApplicationBuilder {
             initial_state: initial_state,
