+++ src/renderer/src/lib.rs
@@ -178,6 +178,7 @@ pub struct Fragment<R: gfx::Resources> {
 }
 
 /// A basic light
+#[derive(Clone)]
 pub struct Light {
     /// The XYZ coordinate of the light
     pub center: [f32; 3],
